+++ libevmasm/SemanticInformation.cpp
@@ -218,6 +218,8 @@ bool SemanticInformation::invalidatesStorage(Instruction _instruction)
 	case Instruction::SENDASSET:
 	case Instruction::ISDELEGATE:
 	case Instruction::GETDELEGATEINFO:
+	case Instruction::GETDELEGATEINFO5:
+	case Instruction::GETDELEGATETOTALVOTE:
 		return true;
 	default:
 		return false;
@@ -277,6 +279,8 @@ bool SemanticInformation::invalidInViewFunctions(Instruction _instruction)
 	case Instruction::TRANSFERASSET:
 	case Instruction::ISDELEGATE:
 	case Instruction::GETDELEGATEINFO:
+	case Instruction::GETDELEGATEINFO5:
+	case Instruction::GETDELEGATETOTALVOTE:
 		return true;
 	default:
 		break;