+++ libsolidity/ast/Types.h
@@ -899,6 +899,7 @@ public:
 		SendAsset,
 		IsDelegate, /// < is delegate
 		GetDelegateInfo, /// < get information of delegate
+		GetDelegateInfo5, /// < get information of five delegate 
 		Transfer, ///< CALL, but without data and throws on error
 		SHA3, ///< SHA3
 		Selfdestruct, ///< SELFDESTRUCT